Milestone-1/Voice_Email_Assistant_M1_FACE_LOGIN_WORKING/Voice_Email_Assistant_M1    milestone 1/requirements.txt
Flask
google-auth
google-auth-oauthlib
google-api-python-client
opencv-python
numpy
